:recycle: Move options menu to fragments
