+++ app/src/main/java/com/kyhsgeekcode/disassembler/IOnBackPressed.kt
@@ -1,0 +1,5 @@
+package com.kyhsgeekcode.disassembler
+
+interface IOnBackPressed {
+    fun onBackPressed(): Boolean
+}
